Removed unnecessary imports, changed package structure.
