+++ src/main/java/de/uhd/ifi/se/decision/management/jira/classification/preprocessing/PreprocessorImpl.java
@@ -1,7 +1,5 @@
-package de.uhd.ifi.se.decision.management.jira.extraction.impl;
+package de.uhd.ifi.se.decision.management.jira.classification.preprocessing;
 
-import de.uhd.ifi.se.decision.management.jira.extraction.Preprocessor;
-import gnu.trove.impl.sync.TSynchronizedShortObjectMap;
 import opennlp.tools.lemmatizer.DictionaryLemmatizer;
 import opennlp.tools.lemmatizer.Lemmatizer;
 import opennlp.tools.namefind.NameFinderME;
@@ -13,11 +11,7 @@ import opennlp.tools.tokenize.TokenizerME;
 import opennlp.tools.tokenize.TokenizerModel;
 
 import java.io.*;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.*;
-import java.util.stream.Collectors;
 
 public class PreprocessorImpl implements Preprocessor {
 
@@ -102,7 +96,7 @@ public class PreprocessorImpl implements Preprocessor {
     @Override
     public List convertToNumbers(List<String> tokens) {
         List<List<Double>> numberTokens = new ArrayList<>();
-        PreTrainedGlove glove = PreTrainedGlove.getInstance();
+        PreTrainedGloveSingleton glove = PreTrainedGloveSingleton.getInstance();
         for (String wordToken : tokens) {
             numberTokens.add(glove.getWordVector(wordToken));
         }
@@ -137,70 +131,6 @@ public class PreprocessorImpl implements Preprocessor {
         List<List<Double>> numberTokens = this.convertToNumbers(tokens);
 
         return this.generateNGram(numberTokens, this.nGramN);
-    }
-}
-
-class PreTrainedGlove {
-
-    private static PreTrainedGlove instance;
-    private Map<String, Double[]> map;
-    private Integer dimensions;
-    private static String GLOVE_FILE_PATH = "src" + File.separator + "main" + File.separator + "resources"
-            + File.separator + "classifier" + File.separator
-            + "language_models" + File.separator;
-    ;
-
-    private PreTrainedGlove(Integer dimensions) {
-        this.dimensions = dimensions;
-        this.map = new HashMap<>();
-        String fullFilename = GLOVE_FILE_PATH + "glove.6b." + dimensions + "d.csv";
-        try (BufferedReader br = Files.newBufferedReader(Paths.get(fullFilename),
-                StandardCharsets.UTF_8)) {
-            String line = br.readLine();
-            while (line != null) {
-                String[] attributes = line.split(" ");
-
-                List<Double> vector = new ArrayList<Double>();
-                //Skip first entry because that is the word.
-                for (int i = 1; i < attributes.length; i++) {
-                    vector.add(Double.parseDouble(attributes[i]));
-                }
-                map.put(attributes[0], Arrays.copyOf(vector.toArray(), vector.size(), Double[].class));
-                line = br.readLine();
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-    }
-
-
-    public static PreTrainedGlove getInstance() {
-        return PreTrainedGlove.getInstance(50);
-    }
-
-    // This method is private because at the moment only the 50D vector is used.
-    private static PreTrainedGlove getInstance(Integer dimensions) {
-        if (PreTrainedGlove.instance == null) {
-            PreTrainedGlove.instance = new PreTrainedGlove(dimensions);
-        }
-        return PreTrainedGlove.instance;
-    }
-
-    /**
-     *  This method gets a word as a parameter and returns a List of Double values representing the relationship between words.
-     *  If no relationship status is known a Lost of zeroes is returned.
-     *
-     * @param word holds the string for which a vector has to be determined
-     * @return
-     */
-    List<Double> getWordVector(String word) {
-        try{
-            return Arrays.asList(this.map.get(word));
-        } catch (Exception e){
-            return new ArrayList<Double>(Collections.nCopies(this.dimensions, 0.0));
-        }
-
     }
 }
