--- src/main/java/de/uhd/ifi/se/decision/management/jira/extraction/ClassificationTrainerCSV.java
@@ -1,4 +1,0 @@
-package de.uhd.ifi.se.decision.management.jira.extraction;
-
-public abstract class ClassificationTrainerCSV implements ClassificationTrainer {
-}
